split brightness and departures into two services
