+++ Service/appSettings.json
@@ -1,21 +1,22 @@
 {
   "LedUpdaterService": {
-    "SignUpdateInterval": 6000,
+    "BrightnessUpdateInterval": 60000, // 1 minute
+    "DarkModeBrightness": 20,
     "LightModeBrightness": 127,
-    "DarkModeBrightness": 20
+    "SignUpdateInterval": 6000 // 6 seconds
   },
   "IpDisplaysApiClient": {
     "TimeoutMiliseconds": 16000
   },
   "SanityApi": {
+    "ApiVersion": "v2024-06-20",
     "Dataset": "production",
-    "UseCdn": true,
-    "ApiVersion": "v2024-06-20"
+    "UseCdn": true
   },
   "RealtimeClientConfig": {
+    "DarkModeUrl": "https://kiosk-api.mtd.org/time/dark-mode",
     "DeparturesUrl": "https://kiosk-api.mtd.org/departures",
-    "GeneralMessagingUrl": "https://kiosk-api.mtd.org/general-messaging",
-    "DarkModeUrl": "https://kiosk-api.mtd.org/time/dark-mode"
+    "GeneralMessagingUrl": "https://kiosk-api.mtd.org/general-messaging"
   },
   "Serilog": {
     "Enrich": [
@@ -54,9 +55,9 @@
       },
       {
         "Args": {
-          "theme": "Serilog.Sinks.SystemConsole.Themes.AnsiConsoleTheme::Code, Serilog.Sinks.Console",
           "outputTemplate": "[{Timestamp:HH:mm:ss} {Level:u3}] {Message:lj} <s:{SourceContext}>{NewLine}{Exception}",
-          "restrictedToMinimumLevel": "Debug"
+          "restrictedToMinimumLevel": "Debug",
+          "theme": "Serilog.Sinks.SystemConsole.Themes.AnsiConsoleTheme::Code, Serilog.Sinks.Console"
         },
         "Name": "Console"
       },
@@ -64,8 +65,8 @@
         "Name": "File",
         "Args": {
           "path": "C:\\Service\\LED_Service\\Log\\log.txt",
-          "rollingInterval": "Day",
-          "retainedFileCountLimit": 7
+          "retainedFileCountLimit": 7,
+          "rollingInterval": "Day"
         }
       }
     ]
